Add support for OpenStack Newton

+++ Cinder/Juno/hypermetro.py
@@ -12,11 +12,11 @@
 #    WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
 #    License for the specific language governing permissions and limitations
 #    under the License.
-#
 
 from cinder import exception
 from cinder.i18n import _, _LI, _LW
 from cinder.openstack.common import log as logging
+from cinder import utils
 from cinder.volume.drivers.huawei import constants
 from cinder.volume.drivers.huawei import huawei_utils
 
@@ -34,8 +34,11 @@ class HuaweiHyperMetro(object):
         """Create hypermetro."""
 
         try:
+            # Check remote metro domain is valid.
+            domain_id = self._valid_rmt_metro_domain()
+
             # Get the remote pool info.
-            config_pool = self.configuration.metro_storage_pools
+            config_pool = self.rmt_client.storage_pools[0]
             remote_pool = self.rmt_client.get_all_pools()
             pool = self.rmt_client.get_pool_info(config_pool, remote_pool)
             if not pool:
@@ -50,9 +53,8 @@ class HuaweiHyperMetro(object):
 
             # Get hypermetro domain.
             try:
-                domain_name = self.configuration.metro_domain_name
-                domain_id = self.rmt_client.get_hyper_domain_id(domain_name)
-                self._wait_volume_ready(remote_lun_id)
+                self._wait_volume_ready(local_lun_id, True)
+                self._wait_volume_ready(remote_lun_id, False)
                 hypermetro = self._create_hypermetro_pair(domain_id,
                                                           local_lun_id,
                                                           remote_lun_id)
@@ -77,16 +79,16 @@ class HuaweiHyperMetro(object):
         metro_id = metadata['hypermetro_id']
         remote_lun_id = metadata['remote_lun_id']
 
-        if metro_id:
-            self.check_metro_need_to_stop(volume)
-
-            # Delete hypermetro
+        # Delete hypermetro.
+        if metro_id and self.client.check_hypermetro_exist(metro_id):
+            self.check_metro_need_to_stop(metro_id)
             self.client.delete_hypermetro(metro_id)
 
         # Delete remote lun.
         if remote_lun_id and self.rmt_client.check_lun_exist(remote_lun_id):
             self.rmt_client.delete_lun(remote_lun_id)
 
+    @utils.synchronized('huawei_create_hypermetro_pair', external=True)
     def _create_hypermetro_pair(self, domain_id, lun_id, remote_lun_id):
         """Create a HyperMetroPair."""
         hcp_param = {"DOMAINID": domain_id,
@@ -102,7 +104,7 @@ class HuaweiHyperMetro(object):
     def connect_volume_fc(self, volume, connector):
         """Create map between a volume and a host for FC."""
         wwns = connector['wwpns']
-        volume_name = huawei_utils.encode_name(volume['id'])
+        volume_name = huawei_utils.encode_name(volume.id)
 
         LOG.info(_LI(
             'initialize_connection_fc, initiator: %(wwpns)s,'
@@ -132,6 +134,8 @@ class HuaweiHyperMetro(object):
         online_wwns_in_host = (
             self.rmt_client.get_host_online_fc_initiators(host_id))
         online_free_wwns = self.rmt_client.get_online_free_wwns()
+        fc_initiators_on_array = self.rmt_client.get_fc_initiator_on_array()
+        wwns = [i for i in wwns if i in fc_initiators_on_array]
         for wwn in wwns:
             if (wwn not in online_wwns_in_host
                     and wwn not in online_free_wwns):
@@ -171,7 +175,7 @@ class HuaweiHyperMetro(object):
                    'data': {'target_lun': int(host_lun_id),
                             'target_discovered': True,
                             'target_wwn': tgt_port_wwns,
-                            'volume_id': volume['id'],
+                            'volume_id': volume.id,
                             'initiator_target_map': init_targ_map,
                             'map_info': map_info},
                    }
@@ -183,7 +187,7 @@ class HuaweiHyperMetro(object):
     def disconnect_volume_fc(self, volume, connector):
         """Delete map between a volume and a host for FC."""
         wwns = connector['wwpns']
-        volume_name = huawei_utils.encode_name(volume['id'])
+        volume_name = huawei_utils.encode_name(volume.id)
         metadata = huawei_utils.get_volume_metadata(volume)
         lun_id = metadata['remote_lun_id']
         host_name = connector['host']
@@ -234,7 +238,7 @@ class HuaweiHyperMetro(object):
                 lungroup_id = self.rmt_client.find_lungroup_from_map(
                     view_id)
         if lungroup_id:
-            left_lunnum = self.rmt_client.get_lunnum_from_lungroup(
+            left_lunnum = self.rmt_client.get_obj_count_from_lungroup(
                 lungroup_id)
 
         if int(left_lunnum) > 0:
@@ -247,11 +251,12 @@ class HuaweiHyperMetro(object):
 
         return info
 
-    def _wait_volume_ready(self, lun_id):
+    def _wait_volume_ready(self, lun_id, local=True):
         wait_interval = self.configuration.lun_ready_wait_interval
+        client = self.client if local else self.rmt_client
 
         def _volume_ready():
-            result = self.rmt_client.get_lun_info(lun_id)
+            result = client.get_lun_info(lun_id)
             if (result['HEALTHSTATUS'] == constants.STATUS_HEALTH
                and result['RUNNINGSTATUS'] == constants.STATUS_VOLUME_READY):
                 return True
@@ -265,32 +270,29 @@ class HuaweiHyperMetro(object):
         LOG.info(_LI("Create Consistency Group: %(group)s."),
                  {'group': group['id']})
         group_name = huawei_utils.encode_name(group['id'])
-        domain_name = self.configuration.metro_domain_name
-        domain_id = self.client.get_hyper_domain_id(domain_name)
-        if not domain_name or not domain_id:
-            msg = _("The domain_name config in cinder.conf is wrong.")
-            LOG.error(msg)
-            raise exception.VolumeBackendAPIException(data=msg)
-
+        domain_id = self._valid_rmt_metro_domain()
         self.client.create_metrogroup(group_name, group['id'], domain_id)
 
     def delete_consistencygroup(self, context, group, volumes):
         LOG.info(_LI("Delete Consistency Group: %(group)s."),
                  {'group': group['id']})
-        model_update = {}
-        volumes_model_update = []
-        model_update['status'] = group['status']
         metrogroup_id = self.check_consistencygroup_need_to_stop(group)
         if metrogroup_id:
+            # Remove pair from metrogroup.
+            for volume in volumes:
+                metadata = huawei_utils.get_volume_metadata(volume)
+                metro_id = metadata['hypermetro_id']
+                if metro_id and self.client.check_hypermetro_exist(metro_id):
+                    if self._check_metro_in_cg(metro_id, metrogroup_id):
+                        self.client.remove_metro_from_metrogroup(metrogroup_id,
+                                                                 metro_id)
+                else:
+                    err = (_("Hypermetro pair %(id)s doesn't exist on array.")
+                           % {'id': metro_id})
+                    LOG.warning(err)
+
+            # Delete metrogroup.
             self.client.delete_metrogroup(metrogroup_id)
-
-        # Deal with the return volumes info
-        for volume_ref in volumes:
-            volume_update = {'id': volume_ref['id']}
-            volume_update['status'] = 'deleted'
-            volumes_model_update.append(volume_update)
-
-        return model_update, volumes_model_update
 
     def update_consistencygroup(self, context, group,
                                 add_volumes, remove_volumes):
@@ -303,15 +305,32 @@ class HuaweiHyperMetro(object):
         if metrogroup_id:
             # Deal with add volumes to CG
             for volume in add_volumes:
-                metro_id = self.check_metro_need_to_stop(volume)
-                self.client.add_metro_to_metrogroup(metrogroup_id,
-                                                    metro_id)
+                metadata = huawei_utils.get_volume_metadata(volume)
+                metro_id = metadata['hypermetro_id']
+                if metro_id and self.client.check_hypermetro_exist(metro_id):
+                    if not self._check_metro_in_cg(metro_id, metrogroup_id):
+                        self.check_metro_need_to_stop(metro_id)
+                        self.client.add_metro_to_metrogroup(metrogroup_id,
+                                                            metro_id)
+                else:
+                    err_msg = _("Hypermetro pair doesn't exist on array.")
+                    LOG.error(err_msg)
+                    raise exception.VolumeBackendAPIException(data=err_msg)
 
             # Deal with remove volumes from CG
             for volume in remove_volumes:
-                metro_id = self.check_metro_need_to_stop(volume)
-                self.client.remove_metro_from_metrogroup(metrogroup_id,
-                                                         metro_id)
+                metadata = huawei_utils.get_volume_metadata(volume)
+                metro_id = metadata['hypermetro_id']
+                if metro_id and self.client.check_hypermetro_exist(metro_id):
+                    if self._check_metro_in_cg(metro_id, metrogroup_id):
+                        self.check_metro_need_to_stop(metro_id)
+                        self.client.remove_metro_from_metrogroup(metrogroup_id,
+                                                                 metro_id)
+                        self.client.sync_hypermetro(metro_id)
+                else:
+                    err_msg = _("Hypermetro pair doesn't exist on array.")
+                    LOG.error(err_msg)
+                    raise exception.VolumeBackendAPIException(data=err_msg)
 
             new_group_info = self.client.get_metrogroup_by_id(metrogroup_id)
             is_empty = new_group_info["ISEMPTY"]
@@ -324,22 +343,15 @@ class HuaweiHyperMetro(object):
             LOG.error(msg)
             raise exception.VolumeBackendAPIException(data=msg)
 
-    def check_metro_need_to_stop(self, volume):
-        metadata = huawei_utils.get_volume_metadata(volume)
-        metro_id = metadata['hypermetro_id']
-        metro_existed = self.client.check_hypermetro_exist(metro_id)
+    def check_metro_need_to_stop(self, metro_id):
+        metro_info = self.client.get_hypermetro_by_id(metro_id)
+        metro_health_status = metro_info['HEALTHSTATUS']
+        metro_running_status = metro_info['RUNNINGSTATUS']
 
-        if metro_existed:
-            metro_info = self.client.get_hypermetro_by_id(metro_id)
-            metro_health_status = metro_info['HEALTHSTATUS']
-            metro_running_status = metro_info['RUNNINGSTATUS']
-
-            if (metro_health_status == constants.HEALTH_NORMAL and
-                (metro_running_status == constants.RUNNING_NORMAL or
-                    metro_running_status == constants.RUNNING_SYNC)):
-                self.client.stop_hypermetro(metro_id)
-
-        return metro_id
+        if (metro_health_status == constants.HEALTH_NORMAL and
+            (metro_running_status == constants.RUNNING_NORMAL or
+                metro_running_status == constants.RUNNING_SYNC)):
+            self.client.stop_hypermetro(metro_id)
 
     def check_consistencygroup_need_to_stop(self, group):
         group_name = huawei_utils.encode_name(group['id'])
@@ -356,3 +368,23 @@ class HuaweiHyperMetro(object):
                 self.client.stop_metrogroup(metrogroup_id)
 
         return metrogroup_id
+
+    def _check_metro_in_cg(self, metro_id, cg_id):
+        metro_info = self.client.get_hypermetro_by_id(metro_id)
+        if metro_info['ISINCG'] == 'true' and metro_info['CGID'] == cg_id:
+            return True
+        return False
+
+    def _valid_rmt_metro_domain(self):
+        domain_name = self.rmt_client.metro_domain
+        if not domain_name:
+            err_msg = _("Hypermetro domain doesn't config.")
+            LOG.error(err_msg)
+            raise exception.VolumeBackendAPIException(data=err_msg)
+        domain_id = self.rmt_client.get_hyper_domain_id(domain_name)
+        if not domain_id:
+            err_msg = _("Hypermetro domain cannot be found.")
+            LOG.error(err_msg)
+            raise exception.VolumeBackendAPIException(data=err_msg)
+
+        return domain_id
